+++ counting.h
@@ -1,3 +1,3 @@
 #include <lemon/list_graph.h>
 using namespace lemon;
-int triangle_count(const ListGraph& G);
+unsigned long triangle_count(const ListGraph& G);
